backend/requirements.txt
requests
pyyaml
python-dotenv
python-owasp-zap-v2.4
openai
colorama
fastapi
uvicorn[standard]
pydantic
